% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq_date.R
\name{acq_date}
\alias{acq_date}
\title{Return date of S-2 Acquisition}
\usage{
acq_date(acq)
}
\arguments{
\item{acq}{path to the S2 acquisition}
}
\value{
A date string of the S-2 acquisition.
}
\description{
\code{acq_date} returns Date of the Sentinel-2 Acquisition.
}
\examples{

acquisitions <- get_acquisitions("2022", "04", "E:/Grasslands_BioDiv/Data/SatData/")
dt <- acq_date(acquisitions[1])

}
