% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_monthly.R
\name{comp_monthly}
\alias{comp_monthly}
\title{Monthly compositing of Sentinel-2 time series}
\usage{
comp_monthly(df, date.column, stat)
}
\arguments{
\item{df}{data frame (interpolated) with the at plot reflectance.}

\item{date.column}{column name storing the acquisition dates.}

\item{stat}{statistic to use for compositing, e.g. max, min, mean.}
}
\value{
A data frame of monthly S-2 reflectances.
}
\description{
\code{comp_monthly} composites S-2 time series to monthly reflectances.
}
