% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div2sf.R
\name{div2sf}
\alias{div2sf}
\title{Simple Feature of the plot locations with diversity indices.}
\usage{
div2sf(div.df, x.column, y.column, epsg.code, write = F)
}
\arguments{
\item{div.df}{data frame with the diversity indices.}

\item{x.column}{name of the column with x-coordinates.}

\item{y.column}{name of the column with y-coordinates}

\item{epsg.code}{epsg code, e.g. 25832}

\item{write}{True or False, write Geopackage to files.}
}
\value{
An sf object with 4 features, i.e. the plot names, and alpha diversity indices. Geopackage of the div.df is stored in the data-raw folder.
}
\description{
\code{div2sf} returns a simple feature object (sf) with point geometries of the diversity data frame.
}
