% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acquisitions.R
\name{get_acquisitions}
\alias{get_acquisitions}
\title{Get S-2 Acquisition paths}
\usage{
get_acquisitions(year, month, sat_directory)
}
\arguments{
\item{year}{string, indicating the year of the acquisitions}

\item{month}{string, indicating the month, two characters, e.g. "01","02",..., "12"}

\item{sat_directory}{directory where the satellite data is stored}
}
\value{
A character vector of the acquisiton paths for a specific year and month, used to load S-2 masks and bands.
}
\description{
\code{get_acquisitions} returns the acquisiton paths for a specific year and month, used to load S-2 masks and bands
}
\examples{

acquisitions <- get_acquisitions("2022", "04", "E:/Grasslands_BioDiv/Data/SatData/")

}
