% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cover.R
\name{get_cover}
\alias{get_cover}
\title{Percentage of the masked area in the study_area for an acquisition.}
\usage{
get_cover(acq, study_area)
}
\arguments{
\item{acq}{path to the acquisition directory with masks and bands}

\item{study_area}{sf bounding box object of the study area.}
}
\value{
Float, percentage of the masked study area.
}
\description{
\code{get_cover} returns the acquisiton paths for specific year and month, used to load S2 masks and bands
}
