% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_masks.R
\name{get_masks}
\alias{get_masks}
\title{Get Mask for S-2 Acquisition}
\usage{
get_masks(acq)
}
\arguments{
\item{acq}{path to the S2 acquisition folder.}
}
\value{
A character vector of the paths to the two masks for the 10m and 20m bands.
}
\description{
\code{get_masks} returns the Sentinel-2 masks to mask clouds etc. from the satellite image.
}
\examples{

acquisitions <- get_acquisitions("2022", "04", "E:/Grasslands_BioDiv/Data/SatData/")
mask_path <- get_masks(acquisitions[1]) # select one acquisition from the vector

}
