% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_area.R
\name{get_study_area}
\alias{get_study_area}
\title{Returns a bounding box of the study area}
\usage{
get_study_area(df.coords, x.column, y.column)
}
\arguments{
\item{df.coords}{data.frame with the plot coordinates}

\item{x.column}{name of the column with x-coordinates}

\item{y.column}{name of the column with y-coordinates}
}
\value{
A bounding box (sf object) of the study area.
}
\description{
\code{get_study_area} gives the study area to crop the Sentinel-2 bands, coordinates should be in Gauss-Kruger Projection (epsg code: ...)
}
\examples{
df <- readxl::read_excel(path = "E:/Grasslands_BioDiv/Data/Field_Data/SUSALPS_samplingData_BT-RB-FE_2022-2023.xlsx")
study_area <- get_study_area(df, "PlotCenter_x_coord", "PlotCenter_y_coord")

}
