% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_index.R
\name{get_train_index}
\alias{get_train_index}
\title{Partitioning the Data into training and testing split.}
\usage{
get_train_index(rf_data, s)
}
\arguments{
\item{rf_data}{data frame, preprocessed random forest data.}

\item{s}{integer, seed.}
}
\value{
A vector of the data frame indices of the training split, which are 70\% of the samples.
}
\description{
\code{get_train_index} returns the data frame indices of the training split. By default a split of 70:30 is created.
}
