% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.ts.R
\name{interpolate.ts}
\alias{interpolate.ts}
\title{Interpolate S-2 Time Series}
\usage{
interpolate.ts(df, plot.column)
}
\arguments{
\item{df}{dataframe with the plot reflectance.}

\item{plot.column}{name of the column storing the plot names.}
}
\value{
A dataframe with columns S-2 bands and acquisition date, with interpolated NA values.
}
\description{
\code{interpolate.ts} interpolates missing values in the dataframe.
}
