% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_comp_months.R
\name{list_comp_months}
\alias{list_comp_months}
\title{List monthly raster composites}
\usage{
list_comp_months(path, ms)
}
\arguments{
\item{path}{string, indicating the year of the acquisitions}

\item{ms}{vector, indicating the month, e.g. c(3:9) for months March to September.}
}
\value{
A character vector of the paths to the calculated composites.
Code to calculate the composites is provided in the \code{data-raw} folder in the \code{comp.S2.bands.R} script. On request, we can make these composites available.
}
\description{
\code{list_comp_months} returns the monthly raster composites for the spatial predicition.
}
