% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_band.R
\name{load_band}
\alias{load_band}
\title{Load masked Sentinel-2 Band from one acquisition}
\usage{
load_band(band, acq, m, study_area)
}
\arguments{
\item{band}{Sentinel-2 band (character), e.g. "B3".}

\item{acq}{path to the S-2 acquisition}

\item{m}{SpatialRaster of corresponding mask in matching spatial resolution}

\item{study_area}{sf object of bounding box of study area}
}
\value{
SpatRaster of the S-2 reflectances in the study area.
}
\description{
\code{load_band} returns the Sentinel-2 band, cloud masked and with the reflectance as the pixel values.
}
