% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mask.R
\name{load_mask}
\alias{load_mask}
\title{Load Mask for S-2 Acquisition}
\usage{
load_mask(res, mask_path, study_area)
}
\arguments{
\item{res}{resolution of the S-2 band which needs to be masked.}

\item{mask_path}{path to S-2 masks}

\item{study_area}{sf object of bounding box of study area}
}
\value{
A SpatialRaster, terra object with true and false.
}
\description{
\code{load_mask} returns the Sentinel-2 masks to mask clouds etc. from the satellite image.
}
