% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.df.R
\name{pivot.df}
\alias{pivot.df}
\title{Pivots data frame to wide table format needed for Random Forest.}
\usage{
pivot.df(df, days_since_cut = F)
}
\arguments{
\item{df}{dataframe (interpolated) with the monthly plot reflectance.}

\item{days_since_cut}{predictor days_since_cut to be included (T/F)}
}
\value{
A data frame of monthly S-2 reflectances.
}
\description{
\code{pivot.df} pivots the monthly S-2 data frame.
}
