% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.band_composite.R
\name{plt.band_composite}
\alias{plt.band_composite}
\title{Plot Band Combination}
\usage{
plt.band_composite(acq, bands, study_area, df = NULL, add.plots = F)
}
\arguments{
\item{acq}{path to the acquisition directory with masks and bands}

\item{bands}{vector of bands, for RGB e.g.: c("B2", "B3", "B4")}

\item{study_area}{sf bounding box object of the study area.}

\item{df}{Data frame with plot coordinates, default column names are "X" and "Y", points will be added, when add.plots = T}

\item{add.plots}{default is F, if T, points of plot locations will be added.}
}
\value{
A map of the study area, optionally with the plot locations.
}
\description{
\code{plt.band_composite} creates a plot of a specific Sentinel-2 Band Combination.
}
\examples{
bands <- c("B2", "B3", "B4")
acq <- get_acquisitions("2022", "04", "E:/Grasslands_Biodiv/Data/SatData/")[1]
plt.band_composite(acq, bands)

}
