% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.varimp.caret.R
\name{plt.varimp.caret}
\alias{plt.varimp.caret}
\title{Barplot of the 15 most important predictors}
\usage{
plt.varimp.caret(imp.df)
}
\arguments{
\item{imp.df}{data frame, with variables and their relative importance.}
}
\value{
A ggplot barplot with the variables on the y-axis and the importance on the x-axis.
}
\description{
\code{plt.varimp.caret} gives the 15 most important predictors of the random forest model.
}
