% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.varimp.R
\name{plt.varimp}
\alias{plt.varimp}
\title{Summary of the variable importance}
\usage{
plt.varimp(forest)
}
\arguments{
\item{forest}{Ranom Forest Train Object.}
}
\value{
A ggplot with four subplots.
}
\description{
\code{plt.varimp} summarizes the variable importance in four plots, by grouping the variables and suming the importance
by A) Band, B) Year, D) Month and by visualizing the 15 most important variables.
}
