% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{RF}
\alias{RF}
\title{Training the Random Forest model}
\usage{
RF(rf_data, train_index, s)
}
\arguments{
\item{rf_data}{data frame, preprocessed random forest data.}

\item{train_index}{data frame indices of the training split.}

\item{s}{integer, seed.}
}
\value{
A train object, the random forest model to predict alpha diversity from Sentinel-2 reflectances.
}
\description{
\code{RF} returns the Random Forest model trained with the S-2 reflectances and the alpha diversity indices.
A repeated cross validation with 10 folds and 5 repeats is used.
}
