% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_predictors.R
\name{RF_predictors}
\alias{RF_predictors}
\title{Select months to use as predictors for the Random Forest.}
\usage{
RF_predictors(data_frame, m_lst)
}
\arguments{
\item{data_frame}{monthly S-2 reflectance data frame.}

\item{m_lst}{vector of the months to be used in the prediction.}
}
\value{
A data frame.
}
\description{
\code{RF_predictors} slices the monthly Sentinel-2 reflectance data frame to the desired months.
}
