% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_S2_months.R
\name{stack_S2_months}
\alias{stack_S2_months}
\title{Stack monthly raster composites}
\usage{
stack_S2_months(fls)
}
\arguments{
\item{fls}{vector of paths to the monthly raster composites which should be stacked.}
}
\value{
A raster brick object with monthly band composites as layers, e.g. B12_2022.04
}
\description{
\code{stack_S2_months} returns the stack of monthly raster composites for the spatial predicition.
}
