% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.RF.R
\name{summarize.RF}
\alias{summarize.RF}
\title{Scatter Plot of Alpha Diversity Prediction}
\usage{
summarize.RF(
  forest,
  rf_data,
  div_df,
  train_index,
  biodiv_index,
  plot_labels = T
)
}
\arguments{
\item{forest}{Random Forest train object.}

\item{rf_data}{the input data frame to the Random Forest model.}

\item{div_df}{data frame with the alpha diversity indices at plot locations, predictor variables.}

\item{train_index}{data frame indices of the training split.}

\item{biodiv_index}{alpha diversity indice which was used to train and test, either specn (= Species Number),
shannon (= shannon index) or simpson (= simpson index)}

\item{plot_labels}{should points in scatter plot be labled with their plot name (T/F).}
}
\value{
A ggplot scatter plot.
}
\description{
\code{summarize.RF} returns a ggplot scatter plot of the predicted vs. sampled alpha diversity per plot location.
The training and testing splits are visualized as well as the individual R2.
}
