% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.RF.R
\name{write.RF}
\alias{write.RF}
\title{Write Random Forest results to a csv file.}
\usage{
write.RF(pred, biodiv_idx, forest, s, csv.path)
}
\arguments{
\item{pred}{short string to indicate which predictors, compositing metrics etc. were used.}

\item{biodiv_idx}{vector of the months to be used in the prediction.}

\item{forest}{The Random Forest model.}

\item{s}{seed used for the prediction and data partitioning.}

\item{csv.path}{path to the csv file, which should be created or already exists.}
}
\value{
A data frame with the columns: Predictors (= which months were used, which compositing method),
Resp_var (= alpha diversity index used for the prediction), seed, R2, RMSE, stdev, var1, var2, var3 (= the three most important predictors).
}
\description{
\code{write.RF} saves the prediction results from the Random Forest model to a csv file.
With each call of \code{write.RF} the prediction results are appended to the csv file. The csv file serves as a documentation, which predictors etc.
were used with which R2, RMSE and Standard Deviation. Also the three most important predictors are saved.
}
