% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_band.R
\name{load_band}
\alias{load_band}
\title{Load masked Sentinel-2 Band from one acquisition}
\usage{
load_band(band, acq, m, study_area)
}
\arguments{
\item{band}{Sentinel-2 band (character)}

\item{acq}{path to the S2 acquisition}

\item{m}{SpatialRaster of corresponding mask in matching spatial resolution}

\item{study_area}{sf object of bounding box of study area}
}
\value{
A character vector of path to the two masks for the 10m and 20m bands.
}
\description{
\code{load_band} returns the Sentinel-2 masks to mask clouds etc. from the satellite image.
}
