% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_S2_months.R
\name{stack_S2_months}
\alias{stack_S2_months}
\title{Stack monthly raster composites}
\usage{
stack_S2_months(fls, path)
}
\arguments{
\item{fls}{string, indicating the year of the acquisitions}

\item{path}{string, indicating the year of the acquisitions}

\item{filename}{string, indicating the month, two characters, e.g. "01","02",..., "12"}
}
\value{
A character vector of the acquisiton paths for a specific year and month, used to load S2 masks and bands.
}
\description{
\code{stack_S2_months} returns the stack of monthly raster composites for the spatial predicition.
}
\examples{

acquisitions <- get_acquisitions("2022", "04", "E:/Grasslands_BioDiv/Data/SatData/")

}
