% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_to_df.R
\name{wb_to_df}
\alias{wb_to_df}
\title{Convert Excel Workbook to data frame}
\usage{
wb_to_df(wb_path)
}
\arguments{
\item{wb_path}{path to the excel workbook.}
}
\value{
A data frame with S-2 bands and acquisition date as columns.
}
\description{
\code{wb_to_df} converts the excel workbook containing the plot reflectances into a dataframe.
}
